[Raytracing] Add RayId, MissId, and CallableId
These types can be used to pass around names as arguments or
payloads. A potential use case is a path tracer where the TraceRay()
calls are dispatched in a loop in the ray generation shader rather
than recursively from the closest-hit shader. The arguments for the
next TraceRay() call (including the ray type and miss shader) are
passed back from the closest-hit shader.