[GLSL] Add Config.availableExtensions
When generating SPIR-V with Glslang it is not possible
to provide a list of available extensions. Glslang will
assume any known extension to be available. If a required
extension is missing, the generated SPIR-V code may be
invalid for the current platform. Config.availableExtensions
can be used to avoid this scenario by marking extensions
as not supported.